fix(lint): split session-files test suite under the max-lines cap
Heals check-lint on main after #130016.
